Add authorization, validation, and new services
This commit introduces authorization attributes and validation checks across multiple controllers, enhancing security and data integrity. New controllers for auditing, email handling, file uploads, password resets, profiles, and search functionalities have been added. The TokenService now supports token generation, refreshing, and revocation, along with new DTOs for email, file uploads, and pagination. The AppDbContext has been updated with new entities, and the Program.cs file has been modified to include JWT authentication and improved Swagger documentation. Additionally, a GlobalExceptionMiddleware has been implemented for consistent error handling, and various services have been created to encapsulate business logic.

+++ school_api/Services/StudentService.cs
@@ -46,7 +46,7 @@ namespace school_api.Services
                 Email = dto.Email,
                 PhoneNumber = dto.PhoneNumber,
                 UserId = dto.UserId,
-                ParentId = dto.ParentId,
+                ParentId = (int)dto.ParentId,
                 StudentClassId = dto.StudentClassId,
             };
 
@@ -89,6 +89,9 @@ namespace school_api.Services
 
             return true;
         }
+
+
+
 
     }
 }
